[iOS] Embed video views within parent views
There are some cases where react native's dynamic resizing of the views
directly causes problems.  Wrapping the views within plain UIViews
seems to solve the problem.

Additionally, this allows us to set the contentMode on the video views
to AspectFill which will replicate the behavior that was recently
merged into android.