getStats for iOS
This is an implementation of getStats on iOS that mostly conforms to the interface initially proposed in #84 and should conform entirely to #91.  I've tried, however, to

a.) Clean up the code quite a bit with helper functions.
b.) Omit the constraints related changes (I feel that we can land that separately without issue).