Support Twilio Video 2.2 on iOS
This PR updates the library to support version 2.2 of the Twilio Video library.  It does not yet expose any new functionality per-se of v2 of the library; the goal here is feature parity with minimal changes for existing users of the library.

The major user-visible changes are:
* onParticipant{Enabled/Disabled}Track are now broken out into separate events for Audio and Video.
* The track identifier prop type now contains two fields: `participantSid` and `videoTrackSid`.  Generally under the hood we try to use sids everywhere as the idenfitier for participants and tracks.  This would necessitate a small change in the example, namely in the `onParticipantAddedVideoTrack` handler set:

```
this.setState({
  participant,
  track,
});
```

and then pass this in as follows:

```
<TwilioVideoParticipantView trackIdentifier={{ participantSid: this.state.participant.sid, videoTrackSid: this.state.track.trackSid }} />
```