Handle case where audioFocusRequest is NULL (#195)
* Handle case where audioFocusRequest is NULL

In our Sentry logging, we've occasionally, seen an error thrown that is `Illegal audioFocusRequest is null`. We haven't been able to reproduce, but believe that this is a race condition in our code that allows `disconnect` to be called before `connect` has been completed. We've been able to fix but adding guardrails in for this case, but figure that this change might also be helpful to prevent others from having same issue.

* Add secondary check for localParticipant being NULL