Update docs to reflect enable/disable methods in version 2.0 (#197)
These docs threw me off because I wrote my code using `onParticipantDisabledTrack` instead of `onParticipantDisabledAudioTrack` and could not figure out why it was not working as intended. A better fix might be to also raise error or warning when using deprecated methods. But for now, updating the documentation might help future users.